Add event and listener for account deletion and send email

+++ taskmaster/src/main/java/bg/softuni/taskmaster/web/controller/UserRestController.java
@@ -1,9 +1,11 @@
 package bg.softuni.taskmaster.web.controller;
 
+import bg.softuni.taskmaster.events.AccountDeletionEvent;
 import bg.softuni.taskmaster.model.dto.UserInfoDTO;
 import bg.softuni.taskmaster.service.AuthorizationService;
 import bg.softuni.taskmaster.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,6 +16,7 @@ public class UserRestController {
 
     private final UserService userService;
     private final AuthorizationService authorizationService;
+    private final ApplicationEventPublisher publisher;
 
     @GetMapping("/{id}")
     public ResponseEntity<UserInfoDTO> userInfo(@PathVariable Long id) {
@@ -22,7 +25,12 @@ public class UserRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (!userService.exists(id)) {
+            return ResponseEntity.badRequest().build();
+        }
         userService.delete(id);
+        UserInfoDTO info = userService.getInfo(id);
+        publisher.publishEvent(new AccountDeletionEvent(this, info.getUsername(), info.getEmail()));
         return ResponseEntity.ok().build();
     }
